add images as base64 txt files, send them with api calls, rename recipeId in Vote model and other small changes

+++ database/migrations/2022_10_12_072737_create_votes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id("voteId");
             $table->integer("userId");
             $table->integer("recipeId");
-            $table->integer("vote");
+            $table->float("vote");
             $table->timestamps();
         });
     }
